refactor(arch): split the `nrf` arch into multiple submodule files
Split the `nrf` arch to make it more manageable when it keeps growing.
